Add authContext, reset files
