--- src/components/Context.js
@@ -1,15 +1,0 @@
-import React, { useContext } from 'react'
-
-const Context = React.createContext()
-
-export function useAuth(){
-  return useContext(Context)
-}
-
-export function authProvider({children}) {
-  return (
-    <Context.Provider>
-      { children }
-    </Context.Provider>
-  )
-}
